agent-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-ai==0.0.13
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0